utils: move Read related wrappers into dedicate file
Move Read related wrappers into dedicate file.
